Main ContentView renamed and refactored
