+++ NasaAPOP/NasaAPOPApp.swift
@@ -12,7 +12,7 @@ import SwiftUI
 struct NasaAPOPApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            MainView()
         }
     }
